Updated Todo OOP structure

+++ backend/delete_todo.php
@@ -1,7 +1,6 @@
 <?php
 
-require_once "db.php";
-require_once "Todo.php";
+require_once "Loader.php";
 
 if ($_SERVER["REQUEST_METHOD"] === "POST") {
 
@@ -12,7 +11,12 @@ if ($_SERVER["REQUEST_METHOD"] === "POST") {
         exit;
     }
 
-    $todo = new Todo($connection , $id , "" ,"");
+    $todo = Todo::find($id);
+
+    if (!$todo->id) {
+        echo "Task not found";
+        exit;
+    }
 
     $result = $todo->delete();
